RequirementMachine: Split off completion procedure into its own file
